# backend/internal/apispec/openapi.yaml
openapi: 3.0.3
info:
  title: XCreativs Web Platform API
  version: "1.0"
  description: >
    REST API for the XCreativs platform. Bearer (JWT) auth is offered globally via the
    Authorize button; public endpoints ignore the token. Scoped API keys use the
    X-API-Key header. Generated from the live router — do not edit by hand;
    run `go run ./cmd/genopenapi > api/openapi.yaml`.
servers:
  - url: http://localhost:8081
    description: Local
  - url: /
    description: Same origin
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    apiKeyAuth: { type: apiKey, in: header, name: X-API-Key }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
security:
  - bearerAuth: []
paths:
  "/api/v1/admin/analytics":
    get:
      tags: [Admin]
      summary: GET /api/v1/admin/analytics
      operationId: get_api_v1_admin_analytics
      responses:
        '200': { description: Success }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  "/api/v1/admin/applications":
    get:
      tags: [Admin]
      summary: GET /api/v1/admin/applications
      operationId: get_api_v1_admin_applications
      responses:
        '200': { description: Success }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  "/api/v1/admin/applications/{id}/assessments":
    get:
      tags: [Admin]
      summary: GET /api/v1/admin/applications/{id}/assessments
      operationId: get_api_v1_admin_applications_id_assessments
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Success }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [Admin]
      summary: POST /api/v1/admin/applications/{id}/assessments
      operationId: post_api_v1_admin_applications_id_assessments
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: Success }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  "/api/v1/admin/applications/{id}/interviews":
    get:
      tags: [Admin]
      summary: GET /api/v1/admin/applications/{id}/interviews
      operationId: get_api_v1_admin_applications_id_interviews
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Success }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [Admin]
      summary: POST /api/v1/admin/applications/{id}/interviews
      operationId: post_api_v1_admin_applications_id_interviews
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: Success }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  "/api/v1/admin/applications/{id}/status":
    patch:
      tags: [Admin]
      summary: PATCH /api/v1/admin/applications/{id}/status
      operationId: patch_api_v1_admin_applications_id_status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: Success }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  "/api/v1/admin/assessment-challenges":
    get:
      tags: [Admin]
      summary: GET /api/v1/admin/assessment-challenges
      operationId: get_api_v1_admin_assessment_challenges
      responses:
        '200': { description: Success }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [Admin]
      summary: POST /api/v1/admin/assessment-challenges
      operationId: post_api_v1_admin_assessment_challenges
      requestBody:
        required: false
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: Success }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  "/api/v1/admin/assessments/{aid}":
    patch:
      tags: [Admin]
      summary: PATCH /api/v1/admin/assessments/{aid}
      operationId: patch_api_v1_admin_assessments_aid
      parameters:
        - name: aid
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: Success }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  "/api/v1/admin/audio-briefs/{slug}/generate-tts":
    post:
      tags: [Admin]
      summary: POST /api/v1/admin/audio-briefs/{slug}/generate-tts
      operationId: post_api_v1_admin_audio_briefs_slug_generate_tts
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: Success }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  "/api/v1/admin/bookings":
    get:
      tags: [Admin]
      summary: GET /api/v1/admin/bookings
      operationId: get_api_v1_admin_bookings
      responses:
        '200': { description: Success }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  "/api/v1/admin/bookings/{id}":
    put:
      tags: [Admin]
      summary: PUT /api/v1/admin/bookings/{id}
      operationId: put_api_v1_admin_bookings_id
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: Success }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  "/api/v1/admin/concierge/reindex":
    post:
      tags: [Admin]
      summary: POST /api/v1/admin/concierge/reindex
      operationId: post_api_v1_admin_concierge_reindex
      requestBody:
        required: false
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: Success }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  "/api/v1/admin/engagements":
    get:
      tags: [Admin]
      summary: GET /api/v1/admin/engagements
      operationId: get_api_v1_admin_engagements
      responses:
        '200': { description: Success }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [Admin]
      summary: POST /api/v1/admin/engagements
      operationId: post_api_v1_admin_engagements
      requestBody:
        required: false
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: Success }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  "/api/v1/admin/exports/bookings":
    get:
      tags: [Admin]
      summary: GET /api/v1/admin/exports/bookings
      operationId: get_api_v1_admin_exports_bookings
      responses:
        '200': { description: Success }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  "/api/v1/admin/exports/diagnostics":
    get:
      tags: [Admin]
      summary: GET /api/v1/admin/exports/diagnostics
      operationId: get_api_v1_admin_exports_diagnostics
      responses:
        '200': { description: Success }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  "/api/v1/admin/exports/rfps":
    get:
      tags: [Admin]
      summary: GET /api/v1/admin/exports/rfps
      operationId: get_api_v1_admin_exports_rfps
      responses:
        '200': { description: Success }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  "/api/v1/admin/interviews/{iid}":
    patch:
      tags: [Admin]
      summary: PATCH /api/v1/admin/interviews/{iid}
      operationId: patch_api_v1_admin_interviews_iid
      parameters:
        - name: iid
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: Success }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  "/api/v1/admin/notifications/broadcast":
    post:
      tags: [Admin]
      summary: POST /api/v1/admin/notifications/broadcast
      operationId: post_api_v1_admin_notifications_broadcast
      requestBody:
        required: false
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: Success }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  "/api/v1/admin/notifications/send-digests":
    post:
      tags: [Admin]
      summary: POST /api/v1/admin/notifications/send-digests
      operationId: post_api_v1_admin_notifications_send_digests
      requestBody:
        required: false
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: Success }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  "/api/v1/admin/pages":
    get:
      tags: [Admin]
      summary: GET /api/v1/admin/pages
      operationId: get_api_v1_admin_pages
      responses:
        '200': { description: Success }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [Admin]
      summary: POST /api/v1/admin/pages
      operationId: post_api_v1_admin_pages
      requestBody:
        required: false
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: Success }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  "/api/v1/admin/pages/{id}":
    delete:
      tags: [Admin]
      summary: DELETE /api/v1/admin/pages/{id}
      operationId: delete_api_v1_admin_pages_id
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Success }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
    put:
      tags: [Admin]
      summary: PUT /api/v1/admin/pages/{id}
      operationId: put_api_v1_admin_pages_id
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: Success }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  "/api/v1/admin/pages/{slug}":
    get:
      tags: [Admin]
      summary: GET /api/v1/admin/pages/{slug}
      operationId: get_api_v1_admin_pages_slug
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Success }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  "/api/v1/admin/partner/applications":
    get:
      tags: [Admin]
      summary: GET /api/v1/admin/partner/applications
      operationId: get_api_v1_admin_partner_applications
      responses:
        '200': { description: Success }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  "/api/v1/admin/partner/applications/{id}":
    put:
      tags: [Admin]
      summary: PUT /api/v1/admin/partner/applications/{id}
      operationId: put_api_v1_admin_partner_applications_id
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: Success }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
...